Allow GrxPerItemColor to work on-the-fly.
Simply add grx:saveValueOnFly="true" to the xml preference.
By default is false so no need to add the line if it won't be true.

Note: Save the key on every color movement is a waste, so, only use it for statusbar preferences where the user can see immediately the changes. But don't use it for the panel or other non-visible prefs.